+++ src/Netsphere.Database/Game/PlayerMailEntity.cs
@@ -10,11 +10,11 @@ namespace Netsphere.Database.Game
         public long Id { get; set; }
 
         [Column]
-        public long PlayerId { get; set; }
+        public int PlayerId { get; set; }
         public PlayerEntity Player { get; set; }
 
         [Column]
-        public long SenderPlayerId { get; set; }
+        public int SenderPlayerId { get; set; }
         public PlayerEntity SenderPlayer { get; set; }
 
         [Column]
